change ingredient amount data type to double from int

+++ Cookbook/CookModels/Ingredient.cs
@@ -4,7 +4,7 @@
     {
         public Guid Guid { get; set; } = Guid.NewGuid();
         public string Name { get; set; }
-        public int? Quantity { get; set; }
+        public double? Quantity { get; set; }
         public string? Unit { get; set; }
     }
 }
